[Addresses #49798515]Added: models, spec, partial batch API implementation
What:
- Added objects to hold the data required to make a batch API call (models)
- Added the first spec_helper and the first *_spec test
- Arranged the requires in click2mail.rb**
- Added first pass configuration functionality

** This was annoying.  Turns out that because the /lib directory is
included Ruby will do requires like: "models/job" but when you need a list
of those files you have to have "lib/**/*.rb" (/lib prefixed).  So you
have to use Dir to get that list and then filter out the /lib and the file
extensions to get it to work.  The other option is to hard code all of the
files in your gem (this approach was taken by Twilio in there gem).